Code refactoring initial import. More to do.
