+++ TRP_Ask_Me/lambda/custom/index.js
@@ -1,6 +1,10 @@
 "use strict";
 const Alexa = require("alexa-sdk"); // import the library
 const https = require('https');
+const axios = require('axios');
+const SPEECH = require('./handlers/speechUtil');
+const startSearchHandlers = require('./handlers/startSearchHandlers');
+
 
 const APP_ID = "amzn1.ask.skill.256760c6-f794-41d1-a173-d347db50e00e";
 
@@ -8,36 +12,31 @@ const APP_ID = "amzn1.ask.skill.256760c6-f794-41d1-a173-d347db50e00e";
 // --------------------------------- Section 1. Data and Text strings  ---------------------------------
 
 const data = [
-    { productCode: "AME", productName: "Africa & Middle East Fund", ticker: "TRAMX", cusip: "77956H740", shareClass: "Investor Class", assetClass: "Equity", coreCategory: "International Equity/Multi-Cap", "price": "$9.07", morningStarRating: "3", portfolioManager: "Oliver Bell", totalNetOfAssets: "$135.5m", investmentObjective: "The fund seeks long-term growth of capital by investing primarily in the common stocks of companies located (or with primary operations) in Africa and the Middle East." },
-    { productCode: "BCG", productName: "Blue Chip Growth Fund", ticker: "TRBCX", cusip: "77954Q106", shareClass: "Investor Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$110.18", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fund seeks to provide long-term capital growth. Income is a secondary objective." },
-    { productCode: "BCA", productName: "Blue Chip Growth Fund - Advisor Class", ticker: "PABGX", cusip: "77954Q205", shareClass: "Advisor Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$108.25", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fund seeks to provide long-term capital growth. Income is a secondary objective." },
-    { productCode: "BCI", productName: "Blue Chip Growth Fund - I Class", ticker: "TBCIX", cusip: "77954Q403", shareClass: "I Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$110.34", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fund seeks to provide long-term capital growth. Income is a secondary objective." },
-    { productCode: "BCR", productName: "Blue Chip Growth Fund - R Class", ticker: "RRBGX", cusip: "77954Q304", shareClass: "R Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$103.91", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fund seeks to provide long-term capital growth. Income is a secondary objective." },
+    { productCode: "AME", productName: "Africa & Middle East Fund", ticker: "TRAMX", cusip: "77956H740", shareClass: "Investor Class", assetClass: "Equity", coreCategory: "International Equity/Multi-Cap", "price": "$9.07", morningStarRating: "3", portfolioManager: "Oliver Bell", totalNetOfAssets: "$135.5m", investmentObjective: "The fundFinder seeks long-term growth of capital by investing primarily in the common stocks of companies located (or with primary operations) in Africa and the Middle East." },
+    { productCode: "BCG", productName: "Blue Chip Growth Fund", ticker: "TRBCX", cusip: "77954Q106", shareClass: "Investor Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$110.18", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fundFinder seeks to provide long-term capital growth. Income is a secondary objective." },
+    { productCode: "BCA", productName: "Blue Chip Growth Fund - Advisor Class", ticker: "PABGX", cusip: "77954Q205", shareClass: "Advisor Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$108.25", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fundFinder seeks to provide long-term capital growth. Income is a secondary objective." },
+    { productCode: "BCI", productName: "Blue Chip Growth Fund - I Class", ticker: "TBCIX", cusip: "77954Q403", shareClass: "I Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$110.34", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fundFinder seeks to provide long-term capital growth. Income is a secondary objective." },
+    { productCode: "BCR", productName: "Blue Chip Growth Fund - R Class", ticker: "RRBGX", cusip: "77954Q304", shareClass: "R Class", assetClass: "Equity", coreCategory: "Equity/Large-Cap", "price": "$103.91", morningStarRating: "5", portfolioManager: "Larry Puglia", totalNetOfAssets: "$63.3b", investmentObjective: "The fundFinder seeks to provide long-term capital growth. Income is a secondary objective." },
     { productCode: "GFA", productName: "Global Allocation Fund - Advisor Class", ticker: "PAFGX", cusip: "87281T202", shareClass: "Advisor Class", assetClass: "Asset Allocation", coreCategory: "/Asset Allocation/Multi-Asset", "price": "$12.40", morningStarRating: "5", portfolioManager: "Charles M. Shriver", totalNetOfAssets: "$688.8m", investmentObjective: "The Fund seeks long-term capital appreciation and income." },
-    { productCode: "R3A", productName: "Retirement 2010 Fund - R Class", ticker: "RRTAX", cusip: "74149P606", shareClass: "R Class", assetClass: "Target Date", coreCategory: "/Target Date/Retirement Date", "price": "$17.13", morningStarRating: "4", portfolioManager: "Jerome Clark", totalNetOfAssets: "$4.1b", investmentObjective: "The fund seeks the highest total return over time consistent with an emphasis on both capital growth and income." },
-    { productCode: "GUA", productName: "Dynamic Global Bond Fund - Advisor Class", ticker: "PAIEX", cusip: "77956H567", shareClass: "Advisor Class", assetClass: "Fixed Income", coreCategory: "/Fixed Income/International", "price": "$9.43", morningStarRating: "1", portfolioManager: "Arif Husain", totalNetOfAssets: "$4.2b", investmentObjective: "The fund seeks high current income." },
-    { productCode: "SPF", productName: "Spectrum International Fund", ticker: "PSILX", cusip: "779906304", shareClass: "Investor Class", assetClass: "Asset Allocation", coreCategory: "/Asset Allocation/Multi-Asset", "price": "$12.60", morningStarRating: "5", portfolioManager: "Charles M. Shriver", totalNetOfAssets: "$1.5b", investmentObjective: "The fund seeks long-term capital appreciation." },
-    { productCode: "EMF", productName: "Emerging Markets Stock Fund", ticker: "PRMSX", cusip: "77956H864", shareClass: "Investor Class", assetClass: "Asset Allocation", coreCategory: "/Asset Allocation/Multi-Asset", "price": "$40.20", morningStarRating: "5", portfolioManager: "Gonzalo Pangaro", totalNetOfAssets: "$13.2b", investmentObjective: "The fund seeks long-term growth of capital through investments primarily in the common stocks of companies located (or with primary operations) in emerging markets." },
-    { productCode: "GCF", productName: "Global Consumer Fund", ticker: "PGLOX", cusip: "77956H344", shareClass: "Investor Class", assetClass: "Equity", coreCategory: "/International Equity/Sector", "price": "$12.21", morningStarRating: "3", portfolioManager: "Jason Nogueira", totalNetOfAssets: "$19.9m", investmentObjective: "The fund seeks to provide long-term capital growth." }
+    { productCode: "R3A", productName: "Retirement 2010 Fund - R Class", ticker: "RRTAX", cusip: "74149P606", shareClass: "R Class", assetClass: "Target Date", coreCategory: "/Target Date/Retirement Date", "price": "$17.13", morningStarRating: "4", portfolioManager: "Jerome Clark", totalNetOfAssets: "$4.1b", investmentObjective: "The fundFinder seeks the highest total return over time consistent with an emphasis on both capital growth and income." },
+    { productCode: "GUA", productName: "Dynamic Global Bond Fund - Advisor Class", ticker: "PAIEX", cusip: "77956H567", shareClass: "Advisor Class", assetClass: "Fixed Income", coreCategory: "/Fixed Income/International", "price": "$9.43", morningStarRating: "1", portfolioManager: "Arif Husain", totalNetOfAssets: "$4.2b", investmentObjective: "The fundFinder seeks high current income." },
+    { productCode: "SPF", productName: "Spectrum International Fund", ticker: "PSILX", cusip: "779906304", shareClass: "Investor Class", assetClass: "Asset Allocation", coreCategory: "/Asset Allocation/Multi-Asset", "price": "$12.60", morningStarRating: "5", portfolioManager: "Charles M. Shriver", totalNetOfAssets: "$1.5b", investmentObjective: "The fundFinder seeks long-term capital appreciation." },
+    { productCode: "EMF", productName: "Emerging Markets Stock Fund", ticker: "PRMSX", cusip: "77956H864", shareClass: "Investor Class", assetClass: "Asset Allocation", coreCategory: "/Asset Allocation/Multi-Asset", "price": "$40.20", morningStarRating: "5", portfolioManager: "Gonzalo Pangaro", totalNetOfAssets: "$13.2b", investmentObjective: "The fundFinder seeks long-term growth of capital through investments primarily in the common stocks of companies located (or with primary operations) in emerging markets." },
+    { productCode: "GCF", productName: "Global Consumer Fund", ticker: "PGLOX", cusip: "77956H344", shareClass: "Investor Class", assetClass: "Equity", coreCategory: "/International Equity/Sector", "price": "$12.21", morningStarRating: "3", portfolioManager: "Jason Nogueira", totalNetOfAssets: "$19.9m", investmentObjective: "The fundFinder seeks to provide long-term capital growth." }
 ];
 
 //======================================================================================================
 //TODO: Replace these text strings to edit the welcome and help messages
 //======================================================================================================
 
-const skillName = "T Rowe Price";
 
-//This is the welcome message for when a user starts the skill without a specific intent.
-const WELCOME_MESSAGE = "Hello, I am Trusty Alexa from " + skillName + ". \n I go by Trusty for short. \n\nYou can start by asking me about a Mutual Fund. "; //+ getGenericHelpMessage(data);
 
 //This is the message a user will hear when they ask Alexa for help in your skill.
 const HELP_MESSAGE = "I can help you find and subscribe to T. Rowe Price Mutual Funds.";
 
-//This is the message a user will hear when they begin a new search
-const NEW_SEARCH_MESSAGE = getGenericHelpMessage(data);
 
-//This is the message a user will hear when they ask Alexa for help while in the SEARCH state
-const SEARCH_STATE_HELP_MESSAGE = getGenericHelpMessage(data);
+
+
 
 //TODO
 const DESCRIPTION_STATE_HELP_MESSAGE = "Here are some things you can say: Tell me more, or give me his or her contact info";
@@ -45,18 +44,14 @@ const DESCRIPTION_STATE_HELP_MESSAGE = "Here are some things you can say: Tell m
 //TODO
 const MULTIPLE_RESULTS_STATE_HELP_MESSAGE = "Sorry, please say the first and last name of the person you'd like to learn more about";
 
-// This is the message use when the decides to end the search
-const SHUTDOWN_MESSAGE = "Goodbye from Trusty. ";
+
 
 //This is the message a user will hear when they try to cancel or stop the skill.
 const EXIT_SKILL_MESSAGE = "We go beyond the numbers. Goodbye.";
 
-const makePlainText = Alexa.utils.TextUtils.makePlainText;
-const makeRichText = Alexa.utils.TextUtils.makeRichText;
-const makeImage = Alexa.utils.ImageUtils.makeImage;
+
 
 const description = 'What would you like to know about today?';
-const imageURL = 'https://static.seekingalpha.com/uploads/2018/10/31/60842-15410397885898802_origin.png'
 
 // =====================================================================================================
 // ------------------------------ Section 2. Skill Code - Intent Handlers  -----------------------------
@@ -64,157 +59,16 @@ const imageURL = 'https://static.seekingalpha.com/uploads/2018/10/31/60842-15410
 // CAUTION: Editing anything below this line might break your skill.
 //======================================================================================================
 
-const states = {
-    SEARCHMODE: "_SEARCHMODE",
-    DESCRIPTION: "_DESCRIPTION",
-    MULTIPLE_RESULTS: "_MULTIPLE_RESULTS",
-    ACTION: "_ACTION",
-};
 
-const newSessionHandlers = {
-    "LaunchRequest": function() {
-        this.handler.state = states.SEARCHMODE;
-        this.response.speak(WELCOME_MESSAGE).listen(getGenericHelpMessage(data));
-        const builder = new Alexa.templateBuilders.BodyTemplate2Builder();
-        const template = builder.setTitle(WELCOME_MESSAGE)
-            .setImage(makeImage(imageURL))
-            .setTextContent(makeRichText('' + description + ''), null, null)
-            .build();
+const states = require('./handlers/ConstStates');
 
-        this.response.renderTemplate(template);
-        this.emit(':responseReady');
-    },
-    "SearchByFundIntent": function () {
-        console.log("SEARCH INTENT");
-        this.handler.state = states.SEARCHMODE;
-        this.emitWithState("SearchByFundIntent");
-    },
-    "TellMeMoreIntent": function () {
-        this.handler.state = states.SEARCHMODE;
-        this.response.speak(WELCOME_MESSAGE).listen(getGenericHelpMessage(data));
-        this.emit(':responseReady');
-    },
-    "TellMeThisIntent": function () {
-        this.handler.state = states.SEARCHMODE;
-        this.emitWithState("SearchByFundIntent");
-    },
-    "SearchByInfoTypeIntent": function () {
-        this.handler.state = states.SEARCHMODE;
-        this.emitWithState("SearchByInfoTypeIntent");
-    },
-    "getSubscribedFunds": function() {
-        this.handler.state = states.ACTION;
-        this.emitWithState("getSubscribedFunds");
-    },
-    "AMAZON.YesIntent": function () {
-        this.response.speak(getGenericHelpMessage(data)).listen(getGenericHelpMessage(data));
-        this.emit(':responseReady');
-    },
-    "AMAZON.NoIntent": function () {
-        this.response.speak(SHUTDOWN_MESSAGE);
-        this.emit(':responseReady');
-    },
-    "AMAZON.RepeatIntent": function () {
-        this.response.speak(HELP_MESSAGE).listen(getGenericHelpMessage(data));
-        this.emit(':responseReady');
-    },
-    "AMAZON.StopIntent": function () {
-        this.response.speak(EXIT_SKILL_MESSAGE);
-        this.emit(':responseReady');
-    },
-    "AMAZON.CancelIntent": function () {
-        this.response.speak(EXIT_SKILL_MESSAGE);
-        this.emit(':responseReady');
-    },
-    "AMAZON.StartOverIntent": function () {
-        this.handler.state = states.SEARCHMODE;
-        var output = "Ok, starting over. " + getGenericHelpMessage(data);
-        this.response.speak(output).listen(output);
-        this.emit(':responseReady');
-    },
-    "AMAZON.HelpIntent": function () {
-        this.response.speak(HELP_MESSAGE + getGenericHelpMessage(data)).listen(getGenericHelpMessage(data));
-        this.emit(':responseReady');
-    },
-    "SessionEndedRequest": function () {
-        this.emit("AMAZON.StopIntent");
-    },
-    "Unhandled": function () {
-        this.handler.state = states.SEARCHMODE;
-        this.emitWithState("SearchByFundIntent");
-    }
-};
-let startSearchHandlers = Alexa.CreateStateHandler(states.SEARCHMODE, {
-    "AMAZON.YesIntent": function () {
-        this.response.speak(NEW_SEARCH_MESSAGE).listen(NEW_SEARCH_MESSAGE);
-        this.emit(':responseReady');
-    },
-    "AMAZON.NoIntent": function () {
-        this.response.speak(SHUTDOWN_MESSAGE);
-        this.emit(':responseReady');
-    },
-    "AMAZON.RepeatIntent": function () {
-        let output;
-        if (this.attributes.lastSearch) {
-            output = this.attributes.lastSearch.lastSpeech;
-            console.log("repeating last speech");
-        }
-        else {
-            output = getGenericHelpMessage(data);
-            console.log("no last speech availble. outputting standard help message.");
-        }
-        this.emit(":ask", output, output);
-    },
-    "SearchByFundIntent": function () {
-        searchByFundIntentHandler.call(this);
-    },
-    "SearchByInfoTypeIntent": function () {
-        searchByInfoTypeIntentHandler.call(this);
-    },
-    "TellMeThisIntent": function () {
-        this.handler.state = states.DESCRIPTION;
-        this.emitWithState("TellMeThisIntent");
-    },
-    "TellMeMoreIntent": function () {
-        this.handler.state = states.DESCRIPTION;
-        this.emitWithState("TellMeMoreIntent");
-    },
-    "AMAZON.HelpIntent": function () {
-        this.response.speak(getGenericHelpMessage(data)).listen(getGenericHelpMessage(data));
-        this.emit(':responseReady');
-    },
-    "AMAZON.StopIntent": function () {
-        this.response.speak(EXIT_SKILL_MESSAGE);
-        this.emit(':responseReady');
-    },
-    "AMAZON.CancelIntent": function () {
-        this.response.speak(EXIT_SKILL_MESSAGE);
-        this.emit(':responseReady');
-    },
-    "AMAZON.StartOverIntent": function () {
-        this.handler.state = states.SEARCHMODE;
-        var output = "Ok, starting over. " + getGenericHelpMessage(data);
-        this.response.speak(output).listen(output);
-        this.emit(':responseReady');
-    },
-    "getSubscribedFunds": function() {
-        this.handler.state = states.ACTION;
-        this.emitWithState("getSubscribedFunds");
-    },
-    "SessionEndedRequest": function () {
-        this.emit("AMAZON.StopIntent");
-    },
-    "Unhandled": function () {
-        console.log("Unhandled intent in startSearchHandlers");
-        this.response.speak(SEARCH_STATE_HELP_MESSAGE).listen(SEARCH_STATE_HELP_MESSAGE);
-        this.emit(':responseReady');
-    }
-});
+
+
 let multipleSearchResultsHandlers = Alexa.CreateStateHandler(states.MULTIPLE_RESULTS, {
 
     "AMAZON.StartOverIntent": function () {
         this.handler.state = states.SEARCHMODE;
-        var output = "Ok, starting over. " + getGenericHelpMessage(data);
+        var output = "Ok, starting over. " + SPEECH.getGenericHelpMessage(data);
         this.response.speak(output).listen(output);
         this.emit(':responseReady');
     },
@@ -280,8 +134,8 @@ let descriptionHandlers = Alexa.CreateStateHandler(states.DESCRIPTION, {
             this.response.speak(speechOutput).listen(repromptSpeech);
         }
         else {
-            speechOutput = getGenericHelpMessage(data);
-            repromptSpeech = getGenericHelpMessage(data);
+            speechOutput = SPEECH.getGenericHelpMessage(data);
+            repromptSpeech = SPEECH.getGenericHelpMessage(data);
             this.handler.state = states.SEARCHMODE;
             this.response.speak(speechOutput).listen(repromptSpeech);
         }
@@ -347,7 +201,7 @@ let descriptionHandlers = Alexa.CreateStateHandler(states.DESCRIPTION, {
     },
     "AMAZON.StartOverIntent": function () {
         this.handler.state = states.SEARCHMODE;
-        var output = "Ok, starting over. " + getGenericHelpMessage(data);
+        var output = "Ok, starting over. " + SPEECH.getGenericHelpMessage(data);
         this.response.speak(output).listen(output);
         this.emit(':responseReady');
     },
@@ -372,12 +226,12 @@ let actionHandlers = Alexa.CreateStateHandler(states.ACTION, {
         let product = this.attributes.lastSearch.results[0];
 
         this.subscribeToFund(product);
-        this.response.speak("You are subscribed to the fund.");
+        this.response.speak("You are subscribed to the fundFinder.");
         this.emit(':responseReady');
     },
     "getSubscribedFunds": function() {
         getSubscribedFunds.call();
-        this.response.speak("You are subscribed to the fund.");
+        this.response.speak("You are subscribed to the fundFinder.");
         this.emit(':responseReady')
     },
     "AMAZON.StopIntent": function () {
@@ -401,7 +255,7 @@ let actionHandlers = Alexa.CreateStateHandler(states.ACTION, {
     },
     "AMAZON.StartOverIntent": function () {
         this.handler.state = states.SEARCHMODE;
-        var output = "Ok, starting over. " + getGenericHelpMessage(data);
+        var output = "Ok, starting over. " + SPEECH.getGenericHelpMessage(data);
         this.response.speak(output).listen(output);
         this.emit(':responseReady');
     },
@@ -468,73 +322,6 @@ function figureOutWhichSlotToSearchBy(productName, productCode, ticker, assetCla
     }
 }
 
-function searchByFundIntentHandler() {
-    var slots = this.event.request.intent.slots;
-    var fundTerms = isSlotValid(this.event.request, "fundTerms");
-
-    if (fundTerms) {
-        var searchQuery = slots.fundTerms.value;
-        console.log("will begin search with  " + slots.fundTerms.value + " in productName");
-        //var searchResults = searchDatabase(data, searchQuery, "productName");
-        var searchResults = searchAemFundDatabase(searchQuery);
-
-        //saving lastSearch results to the current session
-        let lastSearch = this.attributes.lastSearch = searchResults;
-        let output;
-
-        //saving last intent to session attributes
-        this.attributes.lastSearch.lastIntent = "SearchByFundIntent";
-
-        if (searchResults.count > 1) { //multiple results found
-            console.log("Search completed by fund. Multiple fund results were found");
-            let listOfFundsFound = loopThroughArrayOfObjects(lastSearch.results);
-            output = generateSearchResultsMessage(searchQuery, searchResults.results) + listOfFundsFound + ". Which would you like to learn more about?";
-            this.handler.state = states.MULTIPLE_RESULTS; // change state to MULTIPLE_RESULTS
-            this.attributes.lastSearch.lastSpeech = output;
-            this.response.speak(output).listen(output);
-            const template = loopThroughArrayOfFunds(searchResults.results);
-            this.response.renderTemplate(template);
-        } else if (searchResults.count === 1) { //one result found
-            console.log("one match found");
-            this.handler.state = states.DESCRIPTION; // change state to description
-            output = generateSearchResultsMessage(searchQuery, searchResults.results);
-            this.attributes.lastSearch.lastSpeech = output;
-            // this.emit(":ask", generateSearchResultsMessage(searchQuery,searchResults.results));
-            this.response.speak(output).listen(output);
-            const builder = new Alexa.templateBuilders.BodyTemplate2Builder();
-            const template = builder.setTitle(searchResults.results[0].productName)
-                .setTextContent(makeRichText('Morning Star Rating : ' + searchResults.results[0].morningStarRating + '\n' + 'Price : '
-                    + searchResults.results[0].price + '\n' + 'Total Net Assets : ' + searchResults.results[0].totalNetOfAssets + '\n'
-                    + 'CUSIP : ' + searchResults.results[0].cusip + '\n' + 'Ticker : ' + searchResults.results[0].totalNetOfAssets + '\n'
-                    + 'Share Class : ' + searchResults.results[0].shareClass + '\n' + 'Asset Class : ' + searchResults.results[0].assetClass + '\n'
-                    + 'Core Category : ' + searchResults.results[0].coreCategory + '\n' + 'Portfolio Manager : ' + searchResults.results[0].portfolioManager + '\n'
-                    + 'Investment Objective : ' + searchResults.results[0].investmentObjective
-                    + '\n' + output + ''), null, null)
-                .build();
-            this.response.renderTemplate(template);
-        }
-        else {//no match found
-            console.log("no match found");
-            console.log("searchQuery was  = " + searchQuery);
-            console.log("searchResults.results was  = " + searchResults);
-            output = generateSearchResultsMessage(searchQuery, searchResults.results);
-            this.attributes.lastSearch.lastSpeech = output;
-            // this.emit(":ask", generateSearchResultsMessage(searchQuery,searchResults.results));
-            this.response.speak(output).listen(output);
-
-        }
-    }
-    else {
-        console.log("no searchable slot was provided");
-        console.log("searchQuery was  = " + searchQuery);
-        console.log("searchResults.results was  = " + searchResults);
-
-        this.response.speak(generateSearchResultsMessage(searchQuery, false)).listen(generateSearchResultsMessage(searchQuery, false));
-    }
-
-    this.emit(':responseReady');
-
-}
 
 function searchByInfoTypeIntentHandler() {
     var slots = this.event.request.intent.slots;
@@ -608,22 +395,6 @@ function searchByInfoTypeIntentHandler() {
 // =====================================================================================================
 // ------------------------------- Section 3. Generating Speech Messages -------------------------------
 // =====================================================================================================
-
-function generateNextPromptMessage(person, mode) {
-    let infoTypes = ["price", "investment strategy", "net asset value", "morningstar rating", "portfolio manager"];
-    let prompt;
-
-    if (mode === "current") {
-        // if the mode is current, we should give more informaiton about the current contact
-        prompt = ". You can say - tell me more, or  tell me its " + infoTypes[getRandom(0, infoTypes.length - 1)];
-    }
-    //if the mode is general, we should provide general help information
-    else if (mode === "general") {
-        prompt = ". " + getGenericHelpMessage(data);
-    }
-    return prompt;
-}
-
 function generateSendingCardToAlexaAppMessage(person, mode) {
     let sentence = "I have sent " + person.productName + "'s contact card to your Alexa app" + generateNextPromptMessage(person, mode);
     return sentence;
@@ -637,7 +408,7 @@ function generateSearchResultsMessage(searchQuery, results) {
     if (results) {
         switch (true) {
             case (results.length === 0):
-                sentence = "Hmm. I couldn't find " + searchQuery + ". " + getGenericHelpMessage(data);
+                sentence = "Hmm. I couldn't find " + searchQuery + ". " + SPEECH.getGenericHelpMessage(data);
                 break;
             case (results.length === 1):
                 let product = results[0];
@@ -652,18 +423,15 @@ function generateSearchResultsMessage(searchQuery, results) {
         }
     }
     else {
-        sentence = "Sorry, I didn't quite get that. " + getGenericHelpMessage(data);
+        sentence = "Sorry, I didn't quite get that. " + SPEECH.getGenericHelpMessage(data);
     }
     return sentence;
 }
 
-function getGenericHelpMessage(data) {
-    let sentences = ["ask - what is " + getRandomName(data), "say - find an mutual fund " + getRandomFund(data)];
-    return "You can " + sentences[getRandom(0, sentences.length - 1)];
-}
+
 
 function generateSearchHelpMessage(gender) {
-    let sentence = "Sorry, I don't know that. You can ask me - what's BCG fund's price" ;
+    let sentence = "Sorry, I don't know that. You can ask me - what's BCG fundFinder's price" ;
     return sentence;
 }
 
@@ -684,9 +452,10 @@ function generateSpecificInfoMessage(slots, product) {
         infoTypeValue = slots.infoType.value;
     }
 
-    sentence = product.productName + "'s " + infoTypeValue.toLowerCase() + " is - " + person["say" + infoTypeValue.toLowerCase()] + " . Would you like to find another evangelist? " + getGenericHelpMessage(data);
+    sentence = product.productName + "'s " + infoTypeValue.toLowerCase() + " is - " + person["say" + infoTypeValue.toLowerCase()] + " . Would you like to find another evangelist? " + SPEECH.getGenericHelpMessage(data);
     return optimizeForSpeech(sentence);
 }
+
 
 exports.handler = function (event, context, callback) {
     let alexa = Alexa.handler(event, context);
@@ -700,19 +469,6 @@ exports.handler = function (event, context, callback) {
 // =====================================================================================================
 // For more helper functions, visit the Alexa cookbook at https://github.com/alexa/alexa-cookbook
 //======================================================================================================
-
-function getRandom(min, max) {
-    return Math.floor(Math.random() * (max - min + 1) + min);
-}
-
-function getRandomFund(arrayOfStrings) {
-    return arrayOfStrings[getRandom(0, data.length - 1)].productName;
-}
-
-function getRandomName(arrayOfStrings) {
-    let randomNumber = getRandom(0, data.length - 1);
-    return arrayOfStrings[randomNumber].productName;
-}
 
 function titleCase(str) {
     return str.replace(str[0], str[0].toUpperCase());
@@ -879,24 +635,31 @@ function doRequest(url){
 
 async function searchAemFundDatabase(answer) {
     if(answer) {
-        console.log(`ANSWER!!!!!! ${answer}`);
-        // let encodedQuery = encodeURIComponent(answer);
-        // const URL = `https://www.troweprice.com/aem-services/trp/fai/sitesearch/query?query=${encodedQuery}`;
-        // const response = await doRequest(URL);
+
+        async callAem(query)
+        {
+            let encodedQuery = encodeURIComponent(query);
+            const URL = `https://www.troweprice.com/aem-services/trp/fai/sitesearch/query?query=${encodedQuery}`;
+            const response = await doRequest(URL);
+        };
         const bcgURL ='https://www.troweprice.com/financial-intermediary/us/en/investments/mutual-funds/us-products/blue-chip-growth-fund/jcr:content.json';
         console.log(`CALLING PART1! ${bcgURL}`);
-        let resp1 = await doRequest(bcgURL);
-        let productDetailsURL = `https://io9jvz0wni.execute-api.us-east-1.amazonaws.com/demo-stage?productCode=${resp1.productCode}`;
-        console.log(`FINISHED p1 ${JSON.stringify(resp1)}`);
+        try {
+            let resp1 = await doRequest(bcgURL);
+            let productDetailsURL = `https://io9jvz0wni.execute-api.us-east-1.amazonaws.com/demo-stage?productCode=${resp1.productCode}`;
+            console.log(`FINISHED p1 ${JSON.stringify(resp1)}`);
 
-        console.log(`CALLING PART2! ${productDetailsURL}`);
-        let bcgProductCodeResponse = await doRequest(productDetailsURL);
-        console.log(`FINISHED p2 ${JSON.stringify(bcgProductCodeResponse)}`);
+            console.log(`CALLING PART2! ${productDetailsURL}`);
+            let bcgProductCodeResponse = await doRequest(productDetailsURL);
+            console.log(`FINISHED p2 ${JSON.stringify(bcgProductCodeResponse)}`);
+        } catch (error) {
+            console.log ("Error calling to Product Code. ");
+        }
 
-
+        let results = [bcgProductCodeResponse];
         return {
-            count: 1,
-            results: [bcgProductCodeResponse]
+            count: results.length,
+            results: results
         };
     }
 }
