+++ TRP_Ask_Me/lambda/custom/package.json
@@ -1,20 +1,21 @@
 {
-  "name": "TeamLookup",
+  "name": "IIDFundLookup",
   "version": "1.0.0",
-  "description": "A skill to query a database of team members",
+  "description": "A skill to query investments managed by T. Rowe Price",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [
-    "alexa",
-    "skill",
-    "fact"
+    "t rowe price",
+    "institutional",
+    "financial intermediary"
   ],
-  "author": "Amazon.com",
+  "author": "T. Rowe Price",
   "license": "Apache-2.0",
   "dependencies": {
     "alexa-sdk": "^1.0.0",
+    "axios": "^0.19.0",
     "voice-insights-sdk": "^1.2.2"
   }
 }
